api: Split validation-related functions to a separate file
